requirements.txt
Flask==3.1.2
numpy==1.24.4
Pillow==11.3.0
ultralytics==8.3.203
opencv-python==4.6.0.66